+++ cassandra/io/asyncorereactor.py
@@ -98,7 +98,12 @@ class AsyncoreLoop(object):
     def _cleanup(self, thread):
         self._shutdown = True
         log.debug("Waiting for event loop thread to join...")
-        thread.join()
+        thread.join(timeout=1.0)
+        if thread.is_alive():
+            log.warning(
+                "Event loop thread could not be joined, so shutdown may not be clean. "
+                "Please call Cluster.shutdown() to avoid this.")
+
         log.debug("Event loop thread was joined")
 
     def connection_created(self, connection):